+++ examples/cube/color_map.py
@@ -10,14 +10,13 @@ app.simulation.config.shading.color_mode = 1
 app.simulation.config.wireframe.mode = 2
 app.simulation.config.wireframe.width = 2.0
 app.simulation.config.wireframe.color = [1.0, 0.0, 1.0, 1.0]
-app.simulation.config.data.colormap = "gray"
-app.simulation.camera.pos = [10.0, 0.0, 0.0]
-app.simulation.camera.dir = [-1.0, 0.0, 0.0]
+app.simulation.camera.pos = [18.0, 5.0, 10.0]
+app.simulation.camera.look_anchor()
 app.simulation.load_mesh(path="res/cube.obj")
 
+app.simulation.config.data.colormap = "gray"
+app.simulation.config.selection.labels = True
 mesh = app.simulation.bodies[0].mesh
-nface = len(mesh.facets)
-mesh.values = numpy.arange(nface, dtype=float)
-
-while app.running:
-    app.step()
+nf = len(mesh.facets)
+mesh.values = numpy.arange(nf, dtype=float)
+print(mesh.values)